fix: remove site suffix from migrated education pages
