+++ __tests__/pages/staticPages.test.tsx
@@ -3,12 +3,12 @@ import React from "react";
 import AboutRules from "@/app/about/rules/page";
 import CasaBatllo from "@/app/casabatllo/page";
 import Museum from "@/pages/museum";
-import ElementColumns from "@/pages/element_category/columns";
+import ElementColumns from "@/app/element_category/columns/page";
 import MemeLabDistribution from "@/pages/meme-lab/[id]/distribution";
 import AuthorNft6529 from "@/app/author/nft6529/page";
 import BlogArtists from "@/app/blog/a-tale-of-two-artists/page";
 import CapitalFund from "@/app/capital/fund/page";
-import ElementSections from "@/pages/element_category/sections";
+import ElementSections from "@/app/element_category/sections/page";
 import EmmaPlan from "@/pages/emma/plans/[id]";
 
 jest.mock("next/font/google", () => ({
