--- pages/element_category/columns/index.tsx
@@ -1,22 +1,0 @@
-import React from "react";
-import HeaderPlaceholder from "../../../components/header/HeaderPlaceholder";
-import dynamic from "next/dynamic";
-
-const Header = dynamic(() => import("../../../components/header/Header"), {
-  ssr: false,
-  loading: () => <HeaderPlaceholder />,
-});
-
-const IndexPage = () => (
-  <>
-    <div>
-      <title>Redirecting...</title>
-      <meta httpEquiv="refresh" content="0;url=/" />
-      <p>
-        You are being redirected to <a href="/">/</a>
-      </p>
-    </div>
-  </>
-);
-
-export default IndexPage;
